+++ example/lib/widgets/panel.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:genui/genui.dart';
 import 'package:genui_gen/genui_gen.dart';
-import 'package:json_schema_builder/json_schema_builder.dart';
 
 part 'panel.genui.dart';